backend/requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary>=2.9.9,<3.0.0
django-cors-headers>=4.3.1,<4.4.0
pytest>=7.4.0,<7.5.0
pytest-django>=4.5.2,<4.6.0
pytest-cov>=4.1.0,<4.2.0
factory-boy>=3.3.0,<3.4.0 